+++ lib/Eventful/Poro/InstanceMethods.rb
@@ -1,0 +1,15 @@
+# Eventful/Poro/InstanceMethods.rb
+# Eventful::Poro::InstanceMethods
+
+module Eventful
+  module Poro
+    module InstanceMethods
+
+      def initialize(*)
+        super
+        self.class.register(self)
+      end
+
+    end
+  end
+end